aded DTO layer, refactored User and Student into user, aded enrollments and views

+++ Elective_Courses/src/main/java/com/ierdely/elective_courses/repositories/UsersRepository.java
@@ -1,9 +1,11 @@
 package com.ierdely.elective_courses.repositories;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.ierdely.elective_courses.entities.Student;
 import com.ierdely.elective_courses.entities.User;
 
 
@@ -11,5 +13,7 @@ public interface UsersRepository extends JpaRepository<User, Long> {
 	
     User findByUsername(String username);
     
-    Optional<User> findById(Long id);
+    Optional<User> findById(Integer id);
+    
+
 }
